Update OryUI.agc
You can now set the height and width of all items in the list. All items will use this setting. The setting can be added to OryUICreateList or OryUIUpdateList with the parameter itemSize i.e. OryUIUpdateList(list1, "itemSize:100,8;position:0,10")

Also added the divider between each list item back as it was accidentally removed.